Move SettingsUI to a separate prefab
